crs/libs/libAgents/pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "libAgents"
version = "0.1.0"
description = "Deep Search Based AI Agent framework"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aider-chat>=0.83.2",
    "aiofiles>=24.1.0",
    "claude-code-sdk",
    "code-browser-client @ git+ssh://git@github.com/Team-Atlanta/userspace-code-browser.git#subdirectory=python-client",
    "diskcache>=5.6.3",
    "dotenv>=0.9.9",
    "filelock>=3.12.0",
    "google-genai>=1.9.0",
    "litellm>=1.65.1",
    "pocketflow>=0.0.2",
    "psutil>=5.9.0",
    "pydantic>=2.11.1",
    "pytest-asyncio>=0.26.0",
    "tree-sitter==0.24.0", # pin b/c tree-sitter is notorious for not being backward compatible
    "tree-sitter-cpp==0.23.4",
]

[dependency-groups]
dev = []

[tool.setuptools.packages.find]
where = ["."]
include = ["libAgents*"]

[tool.uv.sources]
claude-code-sdk = { git = "ssh://git@github.com/Team-Atlanta/claude-code-sdk-python.git" }
